refactor(etl): split logging out of observability into its own module
