+++ etl-rs/crates/bin/clickhouse-process/src/main.rs
@@ -46,7 +46,7 @@ struct Cli {
 #[tokio::main]
 async fn main() -> Result<()> {
     color_eyre::install()?;
-    let _tracing_guard = etl::observability::init_tracing("clickhouse-process");
+    let _tracing_guard = etl::logging::init_tracing("clickhouse-process");
 
     let metrics_port = std::env::var("METRICS_PORT")
         .ok()
